Добавлен модуль компонента сортировки
